make it a button instead of auto on exit
